+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxAuxData.hs
@@ -29,17 +29,9 @@ module Cardano.Ledger.Shelley.TxAuxData (
 )
 where
 
-import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
 import Cardano.Ledger.Binary (Annotator (..), DecCBOR (..), EncCBOR (..))
 import qualified Cardano.Ledger.Binary.Plain as Plain (ToCBOR)
-import Cardano.Ledger.Core (Era (..), EraTxAuxData (..))
-import Cardano.Ledger.Hashes (
-  EraIndependentTxAuxData,
-  HashAnnotated,
-  SafeHash,
-  SafeToHash (..),
-  hashAnnotated,
- )
+import Cardano.Ledger.Core
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
   Mem,
@@ -55,7 +47,6 @@ import Cardano.Ledger.Metadata (Metadatum (..), validMetadatum)
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Control.DeepSeq (NFData)
 import Data.Map.Strict (Map)
-import Data.Typeable (Proxy (..))
 import Data.Word (Word64)
 import GHC.Generics (Generic)
 import NoThunks.Class (InspectHeapNamed (..), NoThunks (..))
@@ -110,11 +101,6 @@ instance EraTxAuxData ShelleyEra where
 
   validateTxAuxData _ (ShelleyTxAuxData m) = all validMetadatum m
 
-  hashTxAuxData metadata =
-    AuxiliaryDataHash (makeHashWithExplicitProxys index metadata)
-    where
-      index = Proxy @EraIndependentTxAuxData
-
 instance EqRaw (ShelleyTxAuxData era)
 
 instance HashAnnotated (ShelleyTxAuxData era) EraIndependentTxAuxData where
@@ -124,6 +110,7 @@ hashShelleyTxAuxData ::
   ShelleyTxAuxData era ->
   SafeHash EraIndependentTxAuxData
 hashShelleyTxAuxData = hashAnnotated
+{-# DEPRECATED hashShelleyTxAuxData "In favor of `hashAnnotated`" #-}
 
 pattern ShelleyTxAuxData :: forall era. Era era => Map Word64 Metadatum -> ShelleyTxAuxData era
 pattern ShelleyTxAuxData m <-