+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -32,6 +32,7 @@ module Cardano.Ledger.Core (
   EraTxBody (..),
   txIdTxBody,
   EraTxAuxData (..),
+  hashTxAuxData,
   EraTxWits (..),
   EraScript (..),
   hashScript,
@@ -69,7 +70,6 @@ import Cardano.Ledger.Address (
   decompactAddr,
   isBootstrapCompactAddr,
  )
-import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash)
 import Cardano.Ledger.BaseTypes (ProtVer (..))
 import Cardano.Ledger.Binary (
   Annotator,
@@ -199,7 +199,7 @@ class
 
   withdrawalsTxBodyL :: Lens' (TxBody era) Withdrawals
 
-  auxDataHashTxBodyL :: Lens' (TxBody era) (StrictMaybe AuxiliaryDataHash)
+  auxDataHashTxBodyL :: Lens' (TxBody era) (StrictMaybe TxAuxDataHash)
 
   -- | This getter will produce all inputs from the UTxO map that this transaction might
   -- spend, which ones will depend on the validity of the transaction itself. Starting in
@@ -456,9 +456,11 @@ class
   -- preserved. If you need to retain underlying bytes you can use `translateEraThroughCBOR`
   upgradeTxAuxData :: EraTxAuxData (PreviousEra era) => TxAuxData (PreviousEra era) -> TxAuxData era
 
-  hashTxAuxData :: TxAuxData era -> AuxiliaryDataHash
-
   validateTxAuxData :: ProtVer -> TxAuxData era -> Bool
+
+-- | Compute a hash of `TxAuxData`
+hashTxAuxData :: EraTxAuxData era => TxAuxData era -> TxAuxDataHash
+hashTxAuxData = TxAuxDataHash . hashAnnotated
 
 -- | A collection of witnesses in a Tx
 class